Struct massa_sdk::RpcClient

pub struct RpcClient { /* private fields */ }

TODO ask @yvan-sraka

Implementations

impl RpcClient

pub async fn from_url(url: &str) -> RpcClient

Default constructor

pub async fn stop_node(&self) -> RpcResult<()>

Gracefully stop the node.

pub async fn node_sign_message(&self, message: Vec<u8>) -> RpcResult<PubkeySig>

Sign message with node’s key. Returns the public key that signed the message and the signature.

pub async fn add_staking_secret_keys(
&self,
keypairs: Vec<KeyPair>
) -> RpcResult<()>

Add a vector of new keypair for the node to use to stake. No confirmation to expect.

pub async fn remove_staking_addresses(
&self,
addresses: Vec<Address>
) -> RpcResult<()>

Remove a vector of addresses used to stake. No confirmation to expect.

pub async fn get_staking_addresses(&self) -> RpcResult<PreHashSet<Address>>

Return hash-set of staking addresses.

pub async fn node_ban_by_ip(&self, ips: Vec<IpAddr>) -> RpcResult<()>

Bans given ip address(es) No confirmation to expect.

pub async fn node_ban_by_id(&self, ids: Vec<NodeId>) -> RpcResult<()>

Bans given node id(s) No confirmation to expect.

pub async fn node_unban_by_ip(&self, ips: Vec<IpAddr>) -> RpcResult<()>

Unbans given ip address(es) No confirmation to expect.

pub async fn node_unban_by_id(&self, ids: Vec<NodeId>) -> RpcResult<()>

Unbans given node id(s) No confirmation to expect.

pub async fn node_whitelist(&self, ips: Vec<IpAddr>) -> RpcResult<()>

add ips to whitelist create peer if it was unknown

pub async fn node_remove_from_whitelist(&self, ips: Vec<IpAddr>) -> RpcResult<()>

remove IPs from whitelist

pub async fn get_status(&self) -> RpcResult<NodeStatus>

summary of the current state: time, last final blocks (hash, thread, slot, timestamp), clique count, connected nodes count

pub async fn get_operations(
&self,
operation_ids: Vec<OperationId>
) -> RpcResult<Vec<OperationInfo>>

Returns operations information associated to a given list of operations’ IDs.

pub async fn get_endorsements(
&self,
endorsement_ids: Vec<EndorsementId>
) -> RpcResult<Vec<EndorsementInfo>>

get info on endorsements by ids

pub async fn get_block(&self, block_id: BlockId) -> RpcResult<BlockInfo>

Get information on a block given its BlockId

pub async fn get_filtered_sc_output_event(
&self,
filter: EventFilter
) -> RpcResult<Vec<SCOutputEvent>>

Get events emitted by smart contracts with various filters

pub async fn get_addresses(
&self,
addresses: Vec<Address>
) -> RpcResult<Vec<AddressInfo>>

Get info by addresses

pub async fn get_datastore_entries(
&self,
input: Vec<DatastoreEntryInput>
) -> RpcResult<Vec<DatastoreEntryOutput>>

Get datastore entries

pub async fn send_operations(
&self,
operations: Vec<OperationInput>
) -> RpcResult<Vec<OperationId>>

Adds operations to pool. Returns operations that were ok and sent to pool.

pub async fn execute_read_only_bytecode(
&self,
read_only_execution: ReadOnlyBytecodeExecution
) -> RpcResult<ExecuteReadOnlyResponse>

execute read only bytecode

pub async fn execute_read_only_call(
&self,
read_only_execution: ReadOnlyCall
) -> RpcResult<ExecuteReadOnlyResponse>

execute read only SC call