Struct massa_sdk::RpcClient

pub struct RpcClient { /* private fields */ }

Implementations

impl RpcClient

pub async fn from_url(url: &str) -> RpcClient

Default constructor

pub async fn stop_node(&self) -> RpcResult<()>

Gracefully stop the node.

pub async fn node_sign_message(&self, message: Vec<u8>) -> RpcResult<PubkeySig>

Sign message with node’s key. Returns the public key that signed the message and the signature.

pub async fn add_staking_private_keys(
&self,
private_keys: Vec<PrivateKey>
) -> RpcResult<()>

Add a vec of new private keys for the node to use to stake. No confirmation to expect.

pub async fn remove_staking_addresses(
&self,
addresses: Vec<Address>
) -> RpcResult<()>

Remove a vec of addresses used to stake. No confirmation to expect.

pub async fn get_staking_addresses(&self) -> RpcResult<Set<Address>>

Return hashset of staking addresses.

pub async fn ban(&self, ips: Vec<IpAddr>) -> RpcResult<()>

Bans given node id No confirmation to expect.

pub async fn unban(&self, ips: Vec<IpAddr>) -> RpcResult<()>

Unbans given ip addr No confirmation to expect.

pub async fn execute_read_only_request(
&self,
read_only_execution: ReadOnlyExecution
) -> RpcResult<ExecuteReadOnlyResponse>

execute read only bytecode

pub async fn get_status(&self) -> RpcResult<NodeStatus>

summary of the current state: time, last final blocks (hash, thread, slot, timestamp), clique count, connected nodes count

pub async fn get_operations(
&self,
operation_ids: Vec<OperationId>
) -> RpcResult<Vec<OperationInfo>>

Returns operations information associated to a given list of operations’ IDs.

pub async fn get_endorsements(
&self,
endorsement_ids: Vec<EndorsementId>
) -> RpcResult<Vec<EndorsementInfo>>

pub async fn get_block(&self, block_id: BlockId) -> RpcResult<BlockInfo>

Get information on a block given its BlockId

pub async fn get_addresses(
&self,
addresses: Vec<Address>
) -> RpcResult<Vec<AddressInfo>>

pub async fn send_operations(
&self,
operations: Vec<SignedOperation>
) -> RpcResult<Vec<OperationId>>

Adds operations to pool. Returns operations that were ok and sent to pool.